Audio: Fix some type warning
Former-commit-id: 9ed6d64e771e77d03b91060823efb4236739914b [formerly 0efc1b14cd61f3d33fc642dbe4eb6bf05d58ec7e] [formerly dbc079525b48a2efb8a7917b4b376a318f8d5fae [formerly 0116dff3e391e91eb3a7de6f1e8c8029246d4782]]
Former-commit-id: 27a65bcbd5e499dcc741f76a5c0a31bc9ae09e60 [formerly 97ed7176fd1f0f906229d19e68aedc335a7ca420]
Former-commit-id: 1779855a4f20dc4216648aaed59542fd8c6d7bc8

+++ include/Nazara/Audio/SoundStream.hpp
@@ -21,11 +21,11 @@ namespace Nz
 
 			virtual UInt32 GetDuration() const = 0;
 			virtual AudioFormat GetFormat() const = 0;
-			virtual UInt32 GetSampleCount() const = 0;
+			virtual UInt64 GetSampleCount() const = 0;
 			virtual UInt32 GetSampleRate() const = 0;
 
-			virtual unsigned int Read(void* buffer, unsigned int sampleCount) = 0;
-			virtual void Seek(UInt32 offset) = 0;
+			virtual UInt64 Read(void* buffer, UInt64 sampleCount) = 0;
+			virtual void Seek(UInt64 offset) = 0;
 	};
 }
